Some changes to make decoding functions be in more logical places. Moved `decode(::AtomGraph)` from graphnodefeaturization to atomgraph, also added some generic methods and imports/exports.
