+++ src/atoms/atomgraph.jl
@@ -190,6 +190,14 @@ function Base.show(io::IO, ::MIME"text/plain", ag::AtomGraph)
 
 end
 
+function decode(ag::AtomGraph)
+    @assert !(any(isnothing.([ag.featurization, ag.encoded_features])))
+    decoded = decode(ag.featurization, ag.encoded_features)
+    for (k, v) in decoded
+        v["Symbol"] = ag.elements[k]
+    end
+    return decoded
+end
 
 """
     normalized_laplacian(graph)
