+++ src/atoms/atoms.jl
@@ -1,5 +1,7 @@
 module Atoms
 
+import ..ChemistryFeaturization.decode
+
 include("atomgraph.jl")
 export AtomGraph, visualize
