+++ src/featurizations/graphnodefeaturization.jl
@@ -133,12 +133,3 @@ function decode(fzn::GraphNodeFeaturization, encoded::Matrix{<:Real})
     end
     return decoded
 end
-
-function decode(ag::AtomGraph)
-    @assert !(any(isnothing.([ag.featurization, ag.encoded_features])))
-    decoded = decode(ag.featurization, ag.encoded_features)
-    for (k, v) in decoded
-        v["Symbol"] = ag.elements[k]
-    end
-    return decoded
-end
